Initial Commit
Started looking into the content of the ISO for the MC LCE PS3 Edition, and I found some PCK files to test against! So I got a little interested in trying to implement this again. I thought I had looked into it before already, but looks like it was actually GRF that I looked into, not PCK. PCK is what resource packs are stored with, and GRF is what world configuration files are stored with, if I remember right. Say for example like the Tutorial world and it's custom non-vanilla features, like the tutorial menu setup. Those are configurable to run like that differently in-game, it isn't specifically coded as part of the game, it's like an API to add those custom behaviors yourself.

Thanks @PhoenixARC! This will help out a bunch :D
https://github.com/PhoenixARC/LCE_Format_Documentation/blob/main/Formats/PCK.md